Move CreatePaymentCommandValidator.cs
